refactor: pages/Vote -> components로 이동 및 pages, styles 삭제
