+++ apps/web/tsconfig.json
@@ -15,9 +15,7 @@
       "@components/*": ["src/components/*"],
       "@constants/*": ["src/constants/*"],
       "@hooks/*": ["src/hooks/*"],
-      "@utils/*": ["src/utils/*"],
-      "@styles/*": ["src/styles/*"],
-      "@pages/*": ["src/pages/*"]
+      "@utils/*": ["src/utils/*"]
     }
   },
   "include": [
